Move project run tests to projects
